\name{cat1}
\alias{cat1}
\title{One-sample categorical data (using CLT approximation)}
\usage{
  cat1()
}
\description{
Randomly generates a one-sample categorical data problem.  The answers given
are based on the Central Limit Theorem approximation, not the exact answers
based on the binomial distribution.
}
