\name{distr}
\alias{distr}
\title{Distribution questions: normal, binomial, central limit theorem}
\usage{
  distr()
}
\description{
Randomly generates a question about distributions.  Depending on the problem,
you may have to use the normal distribution, the binomial distribution, or
the central limit theorem.  Most of the challenge comes from recognizing which
distribution is appropriate given the circumstances.
}
