\name{quiz3}
\alias{quiz3}
\title{Randomly draws a question that might appear on quiz 3}
\usage{
  quiz3()
}
\details{
  This function essentially just calls \code{distr()} for a question
  about distributions, \code{cat1()} for a one-sample categorical data
  question, or \code{cont1()} for a one-sample continuous data
  question.}
