\name{quiz5}
\alias{quiz5}
\title{Randomly draws a question that might appear on "quiz 5"}
\usage{
  quiz5()
}
\description{
  This function essentially just calls \code{surv()} for a survival analysis
  question or \code{ANOVA()} for an ANOVA question.}
\details{
  There are only 4 quizzes in this class.  These are just problems that weren't
  on any of the first 4 quizzes.}
